+++ lib/resty/validation.lua
@@ -1,5 +1,4 @@
-local _VERSION = "2.5"
-
+local _VERSION = "2.6"
 local setmetatable = setmetatable
 local getmetatable = getmetatable
 local rawget = rawget
@@ -62,6 +61,14 @@ local function istype(t)
         return function(value)
             return t == iotype(value)
         end
+    elseif t == "callable" then
+        return function(value)
+            if type(value) == "function" then
+                return true
+            end
+            local m = getmetatable(value)
+            return m and type(m.__call) == "function"
+        end
     else
         return function(value)
             return t == type(value)
@@ -103,6 +110,9 @@ function factory.userdata()
 end
 function factory.func()
     return factory.type "function"
+end
+function factory.callable()
+    return factory.type "callable"
 end
 factory["function"] = factory.func
 function factory.thread()
@@ -440,6 +450,7 @@ local validators = setmetatable({
     userdata     = factory.userdata(),
     ["function"] = factory.func(),
     func         = factory.func(),
+    callable     = factory.callable(),
     thread       = factory.thread(),
     integer      = factory.integer(),
     float        = factory.float(),
@@ -482,7 +493,7 @@ function data:__call(...)
 end
 local field = {}
 field.__index = field
-function field.new(name, input )
+function field.new(name, input)
     return setmetatable({
         name = name,
         input = input,
@@ -600,8 +611,75 @@ function group:compare(comparison)
                 v1:accept(x)
                 v2:accept(y)
             else
-                v1:reject("compare")
-                v2:reject("compare")
+                v1:reject "compare"
+                v2:reject "compare"
+            end
+        end
+    end
+end
+function group:requisite(r)
+    local c = #r
+    self[#self+1] = function(fields)
+        local n = c
+        local valid = true
+        for i = 1, c do
+            local f = r[i]
+            if not fields[f] then
+                fields[f] = field.new(f)
+            end
+            local field = fields[f]
+            if field.valid then
+                local v = field.value
+                if v == nil or v == "" then
+                    n = n - 1
+                end
+            end
+        end
+        if n > 0 then
+            for i = 1, c do
+                local f = fields[r[i]]
+                if f.valid then
+                    f:accept(f.value)
+                end
+            end
+        else
+            for i = 1, c do
+                local f = fields[r[i]]
+                f:reject "requisite"
+            end
+        end
+    end
+end
+function group:requisites(r, n)
+    local c = #r
+    local n = n or c
+    self[#self+1] = function(fields)
+        local j = c
+        local valid = true
+        for i = 1, c do
+            local f = r[i]
+            if not fields[f] then
+                fields[f] = field.new(f)
+            end
+            local field = fields[f]
+            if field.valid then
+                local v = field.value
+                if v == nil or v == "" then
+                    j = j - 1
+                end
+            end
+        end
+        if n <= j then
+            for i = 1, c do
+                local f = fields[r[i]]
+                if f.valid then
+                    f:accept(f.value)
+                end
+            end
+        else
+            for i = 1, c do
+                local f = fields[r[i]]
+                f:reject "requisites"
             end
         end
     end